Beginning sketch of image processing pipeline.
Just enough of a sketch of our image processing pipeline that we could
get the barest of bare demos running. The rough idea here is to, when a
Blob is created with the tangl.es/code/blobs package, use the Publisher
interface to push that Blob onto the queue, if it's an image blob.
Listener interfaces will pull those Blobs off the queue, and call the
callback. The callback will download the image from the blobs.Storer,
run it through a Processor, and upload it again as a new Blob (directly,
so we don't trigger another queue message).

Our proof of concept of all of this is using GCP's Pub/Sub API to be our
queue, and running images through SQIP to get low-filesize versions we
can embed as placeholders, then upgrade after initial load.

We're going to have to store the headers as JSONB in Postgres, but
that's probably fine, as we're largely treating them as opaque, and as a
way for processors to communicate with our render-time code, so the
renderers can know things like "this is a placeholder image that should
be rendered inline".